+++ frontend/pubspec.yaml
@@ -45,7 +45,7 @@ dependencies:
   flutter_svg: ^2.0.16
   intl: ^0.20.1
   cloud_functions: ^5.2.0
-
+  json_annotation: ^4.9.0
 
 dev_dependencies:
   flutter_test:
@@ -57,6 +57,8 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^4.0.0
+  build_runner: ^2.4.14
+  json_serializable: ^6.9.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
